// lib/types.ts
export type Country = {
  name: string
  slug: string
}

export type City = {
  name: string
  slug: string
}

export type SunriseSunsetData = {
  date: string
  sunrise: string
  sunset: string
  first_light: string
  last_light: string
  dawn: string
  dusk: string
  solar_noon: string
  golden_hour: string
  day_length: string
}